update next js and change paths
