+++ package.json
@@ -15,12 +15,12 @@
     "@types/react-dom": "18.2.4",
     "autoprefixer": "10.4.14",
     "eslint": "8.40.0",
-    "eslint-config-next": "13.4.2",
+    "eslint-config-next": "^13.4.8",
     "format-money-js": "^1.6.3",
-    "next": "13.4.2",
+    "next": "^13.4.8",
     "postcss": "8.4.23",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwindcss": "3.3.2",
     "typescript": "5.0.4"
   },
