+++ src/app/dashboard/(pages)/budgets/page.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+
+export default function Page () {
+
+    return (
+        <section  className='bg-hg-200 w-full h-screen flex flex-col p-10 justify-center items-center'>
+              <h1>Tratamientos</h1>
+        </section>
+      
+    );
+
+};
