+++ src/app/dashboard/(pages)/menu/MenuItems.ts
@@ -3,7 +3,7 @@ export const menuItems = [
       iconSrc: '/images/dashboard/menu-icons/3d_simulator.png',
       altText: 'Simulador 3D',
       title: 'Simulador 3D',
-      link: '/dashboard/pages//crisalix',
+      link: '/dashboard/crisalix',
     },
     {
       iconSrc: '/images/dashboard/menu-icons/budgets.png',
